+++ config/prod.ts
@@ -1,0 +1,11 @@
+import type { UserConfigExport } from '@tarojs/cli'
+
+export default {
+  defineConstants: {
+    CDN_PROTOCOL: '"https"',
+  },
+  mini: {},
+  h5: {
+    publicPath: './',
+  },
+} satisfies UserConfigExport